wiki/requirements.txt
# Core search engine dependencies
datasets>=2.14.0
requests>=2.31.0
tqdm>=4.66.0

# Database and ORM
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Vector search and embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Graph processing and NLP
networkx>=3.0  # For community detection algorithms
spacy>=3.7.0
cdlib>=0.3.0  # Community detection library

# Entity and relationship extraction
gliner>=0.2.0  # GLiNER for entity extraction
gliner-spacy>=0.0.11  # spaCy wrapper for GLiNER
torch>=2.0.0  # Required for GLiNER
transformers>=4.30.0

# Testing
pytest>=7.0.0
pytest-postgresql>=5.0.0  # For database testing

# Development and utilities
python-dotenv>=1.0.0  # Environment variable management
pydantic>=2.0.0  # Data validationglirel==1.2.1
loguru==0.7.3
